bug fix and slots update

+++ src/main/java/qStivi/commands/rpg/Symbol.java
@@ -3,9 +3,9 @@ package qStivi.commands.rpg;
 public class Symbol {
     private final float weight;
     private final String emote;
-    private final int multiplier;
+    private final double multiplier;
 
-    public Symbol(float weight, String emote, int multiplier) {
+    public Symbol(float weight, String emote, double multiplier) {
         this.weight = weight;
         this.emote = emote;
         this.multiplier = multiplier;
@@ -19,7 +19,7 @@ public class Symbol {
         return emote;
     }
 
-    public int getMultiplier() {
+    public double getMultiplier() {
         return multiplier;
     }
 }
